Refactor Editor class.
